refactor(api): consolidate and organize DTOs,
