+++ src/api/EventsApi.ts
@@ -1,49 +1,6 @@
 import { AxiosClient } from '@/core/axios/verbs';
 import { generateQueryParams } from '@/utils/common/api_helper';
-import { Event } from '@/models/Event';
-interface GetEventsPayload {
-	external_customer_id?: string;
-	event_name?: string;
-	start_time?: string;
-	end_time?: string;
-	iter_first_key?: string;
-	iter_last_key?: string;
-	page_size?: number;
-	event_id?: string;
-}
-
-interface GetEventsResponse {
-	events: Event[];
-	has_more: boolean;
-	iter_first_key: string;
-	iter_last_key: string;
-}
-
-interface GetUsageByMeterPayload {
-	meter_id: string;
-	end_time?: string;
-	start_time?: string;
-	external_customer_id?: string;
-	filters?: Record<string, string[]>;
-	window_size?: string;
-}
-
-interface GetUsageByMeterResponse {
-	type: string;
-	event_name: string;
-	results: {
-		window_size: string;
-		value: number;
-	}[];
-}
-
-interface FireEventsPayload {
-	customer_id?: string;
-	subscription_id?: string;
-	feature_id?: string;
-	duration?: number;
-	amount?: number;
-}
+import { GetEventsPayload, GetEventsResponse, GetUsageByMeterPayload, GetUsageByMeterResponse, FireEventsPayload } from '@/types/dto';
 
 class EventsApi {
 	private static baseUrl = '/events';
