+++ src/api/PaymentApi.ts
@@ -1,30 +1,7 @@
 import { AxiosClient } from '@/core/axios/verbs';
-import { PaginationType } from '@/models/Pagination';
 import { Payment } from '@/models/Payment';
 import { generateQueryParams } from '@/utils/common/api_helper';
-
-export interface GetAllPaymentsPayload {
-	currency?: string;
-	destination_id?: string;
-	destination_type?: string;
-	end_time?: string;
-	expand?: string;
-	limit: number;
-	offset: number;
-	order?: 'asc' | 'desc';
-	payment_gateway?: string;
-	payment_ids?: string[];
-	payment_method_type?: string;
-	payment_status?: string;
-	sort?: string;
-	start_time?: string;
-	status?: 'published' | 'deleted' | 'archived' | string;
-}
-
-interface GetAllPaymentsResponse {
-	items: Payment[];
-	pagination: PaginationType;
-}
+import { GetAllPaymentsPayload, GetAllPaymentsResponse } from '@/types/dto';
 
 class PaymentApi {
 	private static baseUrl = '/payments';
