+++ src/pages/product-catalog/plans/PlanDetailsPage.tsx
@@ -5,7 +5,7 @@ import { RouteNames } from '@/core/routes/Routes';
 import { Price } from '@/models/Price';
 import { FeatureType } from '@/models/Feature';
 import { useBreadcrumbsStore } from '@/store/useBreadcrumbsStore';
-import EntitlementApi, { ExtendedEntitlement } from '@/api/EntitlementApi';
+import EntitlementApi from '@/api/EntitlementApi';
 import { PlanApi } from '@/api/PlanApi';
 import formatDate from '@/utils/common/format_date';
 import { getPriceTypeLabel } from '@/utils/common/helper_functions';
@@ -20,8 +20,9 @@ import { getFeatureTypeChips } from '@/components/molecules/CustomerUsageTable/C
 import { formatAmount } from '@/components/atoms/Input/Input';
 import ChargeValueCell from './ChargeValueCell';
 import { BaseEntityStatus } from '@/types/common';
-import { BILLING_PERIOD } from '@/core/data/constants';
-
+import { BILLING_PERIOD } from '@/constants/constants';
+import { Entitlement } from '@/models/Entitlement';
+import { ExtendedEntitlement } from '@/types/dto/Entitlement';
 const formatBillingPeriod = (billingPeriod: string) => {
 	switch (billingPeriod.toUpperCase()) {
 		case BILLING_PERIOD.DAILY:
@@ -90,7 +91,7 @@ const chargeColumns: ColumnData<Price>[] = [
 	},
 ];
 
-const getFeatureValue = (entitlement: ExtendedEntitlement) => {
+const getFeatureValue = (entitlement: Entitlement) => {
 	const value = entitlement.usage_limit?.toFixed() || '';
 
 	switch (entitlement.feature_type) {
