+++ src/types/dto/Meter.ts
@@ -1,0 +1,8 @@
+import { Meter } from '@/models/Meter';
+
+import { PaginationType } from '@/models/Pagination';
+
+export interface getAllMetersResponse {
+	items: Meter[];
+	pagination: PaginationType;
+}
