+++ src/types/dto/Price.ts
@@ -1,0 +1,8 @@
+import { Price } from '@/models/Price';
+
+export interface GetAllPricesResponse {
+	prices: Price[];
+	total: number;
+	offset: number;
+	limit: number;
+}
